using modal for creating post - design change
